disable all sidebars, as on safari, opening the sidebar is very laggy. it makes our app look bad, and we have decided that not many users will need these extra features (the list of sidebars are taken from Control.Sidebar.ts).
since these sidebar buttons are the only ones in the top bar, I have added the insert comment button in place of them.

also disable invert background, as it is associated with dark mode which we have already commented out - see CollaboraOnline/online/issues/9061 (add https://github.com/ infront of it, if i kept the url as is, it would create a link in the collabora repo to this commit).