Rename the File > Repair button to File > Rollback, as I felt it better communicates what it does.
Also fix the english translation files for australian, new zealand, and south african english to use 'initialise' instead of 'initialize', and 'colour' instead of 'color'

I needed to add a new file, ui-en.po in order to override the default name, which is used as the msg's ID (so shouldn't change that). Link to the new file's location in localizations.json.
I couldn't put the msg overrides in the cool-ui.pot file as that is just a template file and it wouldn't change the name.