replace spinner with a gently floating app logo in the loading screen (and make the loading text bigger).
I have also removed the branding.css file I manually added as its not part of the css files listed to build in the makefile. There's already existing logic to get the branding.css file from some config, so I don't want to just simply add it to the list of files to compile as that might break something.

Signed-off-by: Angelina Nguyen <nangelina1412@gmail.com>
Change-Id: If3b455075cb2f32c8b963467c7432bd2e86d33d5